VP-400 Secure org resource for org-admin permission
